+++ src/config/README.md
@@ -1,0 +1,9 @@
+# Sharable Info
+
+Information that is not code, but will be compiled into binary.
+
+1. definitions
+
+2. shared data
+
+3. environmental info, etc.
