+++ src/config/seeds.h
@@ -1,0 +1,9 @@
+#pragma once
+#include <string>
+#include <vector>
+
+namespace p2p
+{
+const std::vector<std::string> seed_nodes_list =
+    {"seeds.moneroseeds.se", "seeds.moneroseeds.ae.org", "seeds.moneroseeds.ch", "seeds.moneroseeds.li"};
+} // namespace p2p
